+++ R/prediction_lm.R
@@ -1,0 +1,3 @@
+#' @rdname prediction
+#' @export
+prediction.lm <- prediction.default