+++ R/prediction_rq.R
@@ -1,0 +1,27 @@
+#' @rdname prediction
+#' @export
+prediction.rq <- 
+function(model, 
+         data = find_data(model, parent.frame()), 
+         type = NULL, 
+         ...) {
+    
+    # extract predicted value at input value
+    if (missing(data)) {
+        pred <- data.frame(fitted = predict(model, ...), se.fitted = NA_real_)
+    } else {
+        pred <- data.frame(fitted = predict(model, newdata = data, ...), 
+                           se.fitted = NA_real_)
+    }
+    names(pred) <- c("fitted", "se.fitted")
+    class(pred[["fitted"]]) <- c("fit", "numeric")
+    class(pred[["se.fitted"]]) <- c("se.fit", "numeric")
+    
+    # obs-x-(ncol(data)+2) data.frame of predictions
+    data <- data
+    structure(if (!length(data)) pred else cbind(data, pred), 
+              class = c("prediction", "data.frame"), 
+              row.names = seq_len(length(pred[["fitted"]])),
+              model.class = class(model),
+              type = type)
+}